Model updates and application
Model gets the original model trained on ALL of the twitter data with .95/.95 accuracies. Model is then applied to reddit comments to get emotions, filtered for those with threshold of at least .6, combined with OG twitter data, and then used to train new model.  The apply file takes the model and applies to some input text.